Removed pages and access because they are not useful as there are only a couple of page types
