+++ src/_templates/default/forms/notebook.jinja2
