--- src/access/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-from access.models import OrganizationalUnit
-
-admin.site.register(OrganizationalUnit)
