--- src/access/urls.py
@@ -1,1 +1,0 @@
-urlpatterns = []
