--- src/access/views.py
@@ -1,1 +1,0 @@
-# from django.shortcuts import render
